+++ examples/futures/FuturesMarketData/GetContractDetails.php
@@ -7,10 +7,15 @@ use BitMart\Lib\CloudConfig;
 require_once __DIR__ . '/../../../vendor/autoload.php';
 
 $APIContract = new APIContractMarket(new CloudConfig([
-    'url' => CloudConst::API_URL_V2_PRO,
+    'url' => "https://demo-api-cloud-v2.bitmart.com",
     'timeoutSecond' => 5,
+    'logger' => [
+        'enabled' => true,
+        'format' => 'default',
+        'outputToConsole' => true,
+    ] 
 ]));
 
-$response = $APIContract->getContractDetails("BTCUSDT")['response'];
+$response = $APIContract->getContractDetails()['response'];
 
 echo json_encode($response);