+++ examples/futures/FuturesMarketData/GetOpenInterest.php
@@ -9,6 +9,11 @@ require_once __DIR__ . '/../../../vendor/autoload.php';
 $APIContract = new APIContractMarket(new CloudConfig([
     'url' => CloudConst::API_URL_V2_PRO,
     'timeoutSecond' => 5,
+    'logger' => [
+        'enabled' => true,
+        'format' => 'default',
+        'outputToConsole' => true,
+    ] 
 ]));;
 
 $response = $APIContract->getContractOpenInterest("BTCUSDT")['response'];