+++ examples/futures/Trading/GetPositionV2.php
@@ -1,21 +1,20 @@
 <?php
 
 use BitMart\Futures\APIContractTrading;
-
 require_once __DIR__ . '/../../ExampleConfig.php';
+
 require_once __DIR__ . '/../../../vendor/autoload.php';
 
 $APIContract = new APIContractTrading(ExampleConfig::getExampleConfig());
 
-$response = $APIContract->getContractPositionRisk(
-)['response'];
+$response = $APIContract->getContractPositionV2()['response'];
 echo json_encode($response);
 
 
-$response = $APIContract->getContractPositionRisk(
+$response = $APIContract->getContractPositionV2(
     [
         'symbol' => 'BTCUSDT',
-        'account' => 'futures',
     ]
 )['response'];
 echo json_encode($response);
+