+++ examples/futures/Trading/GetTransactionHistory.php
@@ -3,24 +3,20 @@
 use BitMart\CloudConst;
 use BitMart\Futures\APIContractTrading;
 use BitMart\Lib\CloudConfig;
-
+require_once __DIR__ . '/../../ExampleConfig.php';
 require_once __DIR__ . '/../../../vendor/autoload.php';
 
-$APIContract = new APIContractTrading(new CloudConfig([
-    'url' => CloudConst::API_URL_V2_PRO,
-    'accessKey' => "<your_api_key>",
-    'secretKey' => "<your_secret_key>",
-    'memo' => "<your_memo>",
-]));
+$APIContract = new APIContractTrading(ExampleConfig::getExampleConfig());
 
 $response = $APIContract->getContractTransactionHistory()['response'];
 echo json_encode($response);
 
 
-$endTime = round(microtime(true))*1000;
-$startTime = $endTime - (60*60)*1000;
+$endTime = round(microtime(true)) * 1000;
+$startTime = $endTime - (60 * 60) * 1000;
 $response = $APIContract->getContractTransactionHistory(
     [
+        'account' => 'futures',
         'flow_type' => 0,
         'start_time' => $startTime,
         'end_time' => $endTime,