+++ examples/futures/Trading/SubmitOrder.php
@@ -1,29 +1,24 @@
 <?php
 
-use BitMart\CloudConst;
 use BitMart\Futures\APIContractTrading;
-use BitMart\Lib\CloudConfig;
 
 require_once __DIR__ . '/../../../vendor/autoload.php';
+require_once __DIR__ . '/../../ExampleConfig.php';
 
-$APIContract = new APIContractTrading(new CloudConfig([
-    'url' => CloudConst::API_URL_V2_PRO,
-    'accessKey' => "<your_api_key>",
-    'secretKey' => "<your_secret_key>",
-    'memo' => "<your_memo>",
-]));
+$APIContract = new APIContractTrading(ExampleConfig::getExampleConfig());
 
 $response = $APIContract->submitOrder(
     'BTCUSDT',
     1,
     [
-        'client_order_id' => "test3000000001",
+        // 'client_order_id' => "test3000000001xx",
         'type' => "limit",
-        'leverage' => "1",
+        'leverage' => "100",
         'open_type' => "isolated",
         'mode' => 1,
-        'price' => "10",
+        'price' => "9200",
         'size' => 1,
+        'stp_mode' => 1,
     ]
 )['response'];